+++ src/_pytest/skipping.py
@@ -12,7 +12,6 @@ import sys
 import traceback
 from typing import Optional
 
-from _pytest._raises_group import AbstractRaises
 from _pytest.config import Config
 from _pytest.config import hookimpl
 from _pytest.config.argparsing import Parser
@@ -21,6 +20,7 @@ from _pytest.nodes import Item
 from _pytest.outcomes import fail
 from _pytest.outcomes import skip
 from _pytest.outcomes import xfail
+from _pytest.raises_group import AbstractRaises
 from _pytest.reports import BaseReport
 from _pytest.reports import TestReport
 from _pytest.runner import CallInfo
